CON-44 (work-in-progress-snapshot)

+++ src/main/scala/io/casperlabs/sim/sim_engine_sequential/AgentRefImpl.scala
@@ -2,5 +2,5 @@ package io.casperlabs.sim.sim_engine_sequential
 
 import io.casperlabs.sim.simulation_framework.AgentRef
 
-case class AgentRefImpl(address: Long) extends AgentRef {
+case class AgentRefImpl(address: Int) extends AgentRef {
 }
